+++ CHANGELOG.md
@@ -11,7 +11,7 @@ contract applies.
 
 ---
 
-## [Unreleased]
+## [0.4.0] - 2026-07-14
 
 ### Added